week-3/requirements.txt
django
django-widget-tweaks
djangorestframework
Markdown # for the Markdown support (pretty printing of browsable API documentations)